modify

+++ modules/myqbapi/src/main/java/allcom/dao/QuestionContentRepository.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Transactional(readOnly = true)
-public interface QuestionContentRepository extends CrudRepository<QuestionContent, Integer> {
+public interface QuestionContentRepository extends CrudRepository<QuestionContent, Long> {
     @Query("from QuestionContent a where a.umid = ?1 and a.content like %?2%")
     List<QuestionContent> findByUmidAndContent(int umid,String content);
 }
